# aidlc-docs/construction/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Opportunity Management Service API
  description: RESTful API for managing opportunities, problem statements, skill requirements, timeline requirements, and attachments
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /problem-statements/{problem_statement_id}/attachments:
    post:
      summary: Add an attachment to a problem statement
      operationId: add_attachment
      parameters:
        - name: problem_statement_id
          in: path
          required: true
          description: Problem statement ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentCreateRequest'
      responses:
        '201':
          description: Attachment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      summary: Get attachments for a problem statement
      operationId: get_attachments_for_problem_statement
      parameters:
        - name: problem_statement_id
          in: path
          required: true
          description: Problem statement ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attachments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /attachments/{attachment_id}:
    delete:
      summary: Remove an attachment
      operationId: remove_attachment
      parameters:
        - name: attachment_id
          in: path
          required: true
          description: Attachment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachmentRemoveRequest'
      responses:
        '200':
          description: Attachment removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /attachments/{attachment_id}/download:
    get:
      summary: Download an attachment file
      operationId: download_attachment
      parameters:
        - name: attachment_id
          in: path
          required: true
          description: Attachment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    StandardResponse:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, RESOURCE_NOT_FOUND, OPERATION_NOT_ALLOWED, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    AttachmentCreateRequest:
      type: object
      required:
        - file
        - uploaded_by
      properties:
        file:
          type: string
          format: binary
        uploaded_by:
          type: string
          format: uuid
    AttachmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        problem_statement_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
        file_url:
          type: string
        uploaded_by:
          type: string
          format: uuid
        uploaded_at:
          type: string
          format: date-time
        is_removed:
          type: boolean
    AttachmentListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                attachments:
                  type: array
                  items:
                    $ref: '#/components/schemas/AttachmentResponse'
    AttachmentRemoveRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'